Modify: move usecase
